Rewriting the orgs state to a slice
