--- src/altinn-app-frontend/src/shared/resources/orgs/fetch/fetchOrgsActionTypes.ts
@@ -1,5 +1,0 @@
-const moduleName = 'ORGS_DATA';
-
-export const FETCH_ORGS = `${moduleName}.FETCH_ORGS`;
-export const FETCH_ORGS_FULFILLED = `${moduleName}.FETCH_ORGS_FULFILLED`;
-export const FETCH_ORGS_REJECTED = `${moduleName}.FETCH_ORGS_REJECTED`;
