+++ src/altinn-app-frontend/src/shared/resources/orgs/index.d.ts
@@ -1,0 +1,14 @@
+import type { IAltinnOrgs } from 'altinn-shared/types';
+
+export interface IOrgsState {
+  allOrgs: IAltinnOrgs;
+  error: Error;
+}
+
+export interface IFetchOrgsFulfilled {
+  orgs: IAltinnOrgs;
+}
+
+export interface IFetchOrgsRejected {
+  error: Error;
+}
