+++ src/altinn-app-frontend/src/shared/resources/queue/appTask/appTaskQueueSagas.test.ts
@@ -4,7 +4,7 @@ import { startInitialAppTaskQueueSaga } from 'src/shared/resources/queue/appTask
 import { startInitialAppTaskQueueFulfilled } from 'src/shared/resources/queue/queueSlice';
 import TextResourcesActions from 'src/shared/resources/textResources/textResourcesActions';
 import { LanguageActions } from 'src/shared/resources/language/languageSlice';
-import OrgsActions from 'src/shared/resources/orgs/orgsActions';
+import { OrgsActions } from 'src/shared/resources/orgs/orgsSlice';
 import { ApplicationMetadataActions } from 'src/shared/resources/applicationMetadata/applicationMetadataSlice';
 import { startInitialUserTaskQueueSaga } from 'src/shared/resources/queue/userTask/userTaskQueueSagas';
 import { FormLayoutActions } from 'src/features/form/layout/formLayoutSlice';
@@ -17,7 +17,7 @@ describe('appTaskQueueSagas', () => {
       .put(LanguageActions.fetchLanguage())
       .put(ApplicationMetadataActions.get())
       .put(FormLayoutActions.fetchLayoutSets())
-      .call(OrgsActions.fetchOrgs)
+      .put(OrgsActions.fetch())
       .put(startInitialAppTaskQueueFulfilled())
       .run();
   });
