+++ src/altinn-app-frontend/src/shared/resources/queue/appTask/appTaskQueueSagas.ts
@@ -5,11 +5,11 @@ import {
   startInitialAppTaskQueueFulfilled,
 } from '../queueSlice';
 import { ApplicationSettingsActions } from '../../applicationSettings/applicationSettingsSlice';
-import OrgsActions from 'src/shared/resources/orgs/orgsActions';
 import TextResourcesActions from '../../textResources/textResourcesActions';
 import { LanguageActions } from '../../language/languageSlice';
 import { FormLayoutActions } from 'src/features/form/layout/formLayoutSlice';
 import { ApplicationMetadataActions } from 'src/shared/resources/applicationMetadata/applicationMetadataSlice';
+import { OrgsActions } from '../../orgs/orgsSlice';
 
 export function* startInitialAppTaskQueueSaga(): SagaIterator {
   yield put(ApplicationSettingsActions.fetchApplicationSettings());
@@ -17,7 +17,7 @@ export function* startInitialAppTaskQueueSaga(): SagaIterator {
   yield put(LanguageActions.fetchLanguage());
   yield put(ApplicationMetadataActions.get());
   yield put(FormLayoutActions.fetchLayoutSets());
-  yield call(OrgsActions.fetchOrgs);
+  yield put(OrgsActions.fetch());
   yield put(startInitialAppTaskQueueFulfilled());
 }
