+++ src/altinn-app-frontend/src/types/index.ts
@@ -13,7 +13,7 @@ import type { IApplicationMetadataState } from 'src/shared/resources/application
 import type { IAttachmentState } from '../shared/resources/attachments';
 import type { IInstanceDataState } from '../shared/resources/instanceData';
 import type { ILanguageState } from '../shared/resources/language/languageSlice';
-import type { IOrgsState } from '../shared/resources/orgs/orgsReducers';
+import type { IOrgsState } from '../shared/resources/orgs';
 import type { IPartyState } from '../shared/resources/party/partyReducers';
 import type { IProcessState } from '../shared/resources/process/processReducer';
 import type { IProfileState } from '../shared/resources/profile/profileReducers';
